requirements.txt
Flask>=2.0
pandas>=1.3
python-dotenv>=0.21
google-generativeai>=0.3.2
matplotlib>=3.5
gunicorn>=20.1
